+++ lib/Map/ImageryProvider/ProtomapsImageryProvider.ts
@@ -93,6 +93,9 @@ interface Options {
   credit?: Credit | string;
   paintRules: PaintRule[];
   labelRules: LabelRule[];
+
+  /** The name of the property that is a unique ID for features */
+  idProperty?: string;
 }
 
 /** Buffer (in pixels) used when rendering (and generating - through geojson-vt) vector tiles */
@@ -284,8 +287,11 @@ export default class ProtomapsImageryProvider
   // Protomaps properties
   /** Data object from constructor options (this is transformed into `source`) */
   private readonly data: ProtomapsData;
+  private readonly maximumNativeZoom: number;
   private readonly labelers: Labelers;
   private readonly view: View | undefined;
+  private idProperty: string;
+
   readonly source: Source;
   readonly paintRules: PaintRule[];
   readonly labelRules: LabelRule[];
@@ -300,6 +306,10 @@ export default class ProtomapsImageryProvider
 
     this.minimumLevel = defaultValue(options.minimumZoom, 0);
     this.maximumLevel = defaultValue(options.maximumZoom, 24);
+    this.maximumNativeZoom = defaultValue(
+      options.maximumNativeZoom,
+      this.maximumLevel
+    );
 
     this.rectangle = isDefined(options.rectangle)
       ? Rectangle.intersection(
@@ -341,6 +351,7 @@ export default class ProtomapsImageryProvider
     // Protomaps
     this.paintRules = options.paintRules;
     this.labelRules = options.labelRules;
+    this.idProperty = options.idProperty ?? "FID";
 
     // Generate protomaps source based on this.data
     // - URL of pmtiles, geojson or pbf files
@@ -348,7 +359,7 @@ export default class ProtomapsImageryProvider
       if (this.data.endsWith(".pmtiles")) {
         this.source = new PmtilesSource(this.data, false);
         let cache = new TileCache(this.source, 1024);
-        this.view = new View(cache, 14, 2);
+        this.view = new View(cache, this.maximumNativeZoom, 2);
       } else if (
         this.data.endsWith(".json") ||
         this.data.endsWith(".geojson")
@@ -357,7 +368,7 @@ export default class ProtomapsImageryProvider
       } else {
         this.source = new ZxySource(this.data, false);
         let cache = new TileCache(this.source, 1024);
-        this.view = new View(cache, 14, 2);
+        this.view = new View(cache, this.maximumNativeZoom, 2);
       }
     }
     // Source object
@@ -638,7 +649,7 @@ export default class ProtomapsImageryProvider
       data,
       minimumZoom: options?.minimumZoom ?? this.minimumLevel,
       maximumZoom: options?.maximumZoom ?? this.maximumLevel,
-      maximumNativeZoom: options?.maximumNativeZoom,
+      maximumNativeZoom: options?.maximumNativeZoom ?? this.maximumNativeZoom,
       rectangle: options?.rectangle ?? this.rectangle,
       credit: options?.credit ?? this.credit,
       paintRules: options?.paintRules ?? this.paintRules,
@@ -660,7 +671,7 @@ export default class ProtomapsImageryProvider
       featureProp = GEOJSON_FEATURE_ID_PROP;
       layerName = GEOJSON_SOURCE_LAYER_NAME;
     } else {
-      featureProp = "FID";
+      featureProp = this.idProperty;
       layerName = feature.properties?.[LAYER_NAME_PROP]?.getValue();
     }