+++ CHANGES.md
@@ -5,6 +5,8 @@
 - Add `aboutButtonHrefUrl` to `configParameters`. Defaults to `"about.html"`. If set to `null`, then the About button will not be shown.
 - Add `refreshIntervalTemplate` to `OpenDataSoftCatalogItemTraits` - this can be used to set `refreshInterval` using Mustache template rendered on ODS Dataset JSON object
 - Add `plugins` property to `ConfigParameters` type
+- Fix `MapboxVectorTileCatalogItem` feature highlighting - this requires use of `idProperty` trait (also added `idProperty` to `ProtomapsImageryProvider`)
+- Add `maximumNativeZoom` to `ProtomapsImageryProvider`
 - [The next improvement]
 
 #### 8.2.20 - 2022-10-20